bookmaker/requirements.txt
Django>=3.2
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
qrcode>=7.0
Pillow>=9.0
bip-utils>=2.7.0
pycountry>=22.0
playwright>=1.30
mnemonic>=0.20
bip32utils>=0.3
python-dotenv>=1.0.0
tqdm>=4.65.0
django-q2>=1.6.1
whitenoise>=6.5.0
gunicorn>=21.2.0
python-dateutil>=2.8.2

